+++ CHANGELOG.md
@@ -1,8 +1,10 @@
+### 1.1.0
 * Fix finder module available db types detection
 * Add query placeholders support for finders
 * Update orient (1.7.8 > 1.7.9) - [important hotfix](https://groups.google.com/forum/#!topic/orient-database/vPF85I5Blts)
 * Add Optional support as finder return type (jdk or guava Optional)
-
+* Add orient module option to disable database auto creation
+* Update guice (4.beta4 -> 4.beta5)
 
 ### 1.0.3 (2014-09-15)